refactor: rename files to index
